+++ addons/event/controllers/__init__.py
@@ -1,7 +1,4 @@
 # -*- coding: utf-8 -*-
 # Part of Odoo. See LICENSE file for full copyright and licensing details.
 
-from . import controllers
-from . import models
-from . import wizard
-from . import report
+from . import main
